extracted class - argumentsParser
