+++ src/main/scala/beam/sim/BeamHelper.scala
@@ -2,8 +2,13 @@ package beam.sim
 
 import java.io.{FileOutputStream, FileWriter}
 import java.nio.file.{Files, Paths, StandardCopyOption}
-import java.util.concurrent.TimeUnit
 import java.util.{Properties, Random}
+import java.util.concurrent.TimeUnit
+
+import scala.collection.JavaConverters._
+import scala.collection.mutable
+import scala.collection.mutable.ListBuffer
+import scala.concurrent.Await
 
 import beam.agentsim.agents.ridehail.{RideHailIterationHistory, RideHailSurgePricingManager}
 import beam.agentsim.events.handling.BeamEventsHandling
@@ -11,19 +16,20 @@ import beam.analysis.ActivityLocationPlotter
 import beam.analysis.plots.{GraphSurgePricing, RideHailRevenueAnalysis}
 import beam.replanning._
 import beam.replanning.utilitybased.UtilityBasedModeChoice
+import beam.router.{BeamSkimmer, RouteHistory, TravelTimeObserved}
 import beam.router.osm.TollCalculator
 import beam.router.r5.{DefaultNetworkCoordinator, FrequencyAdjustingNetworkCoordinator, NetworkCoordinator}
-import beam.router.{BeamSkimmer, RouteHistory, TravelTimeObserved}
 import beam.scoring.BeamScoringFunctionFactory
 import beam.sim.common.GeoUtils
 import beam.sim.config.{BeamConfig, ConfigModule, MatSimBeamConfigBuilder}
 import beam.sim.metrics.Metrics._
 import beam.sim.modules.{BeamAgentModule, UtilsModule}
 import beam.sim.population.PopulationAdjustment
+import beam.sim.ArgumentsParser.{Arguments, Worker}
+import beam.utils.{NetworkHelper, _}
+import beam.utils.scenario.{InputType, ScenarioLoader, ScenarioSource}
 import beam.utils.scenario.matsim.MatsimScenarioSource
 import beam.utils.scenario.urbansim.{CsvScenarioReader, ParquetScenarioReader, UrbanSimScenarioSource}
-import beam.utils.scenario.{InputType, ScenarioLoader, ScenarioSource}
-import beam.utils.{NetworkHelper, _}
 import com.conveyal.r5.transit.TransportNetwork
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.scala.DefaultScalaModule
@@ -31,8 +37,8 @@ import com.google.inject
 import com.typesafe.config.{ConfigFactory, Config => TypesafeConfig}
 import com.typesafe.scalalogging.LazyLogging
 import kamon.Kamon
-import org.matsim.api.core.v01.population.Person
 import org.matsim.api.core.v01.{Id, Scenario}
+import org.matsim.api.core.v01.population.Person
 import org.matsim.core.api.experimental.events.EventsManager
 import org.matsim.core.config.{Config => MatsimConfig}
 import org.matsim.core.config.groups.TravelTimeCalculatorConfigGroup
@@ -43,11 +49,6 @@ import org.matsim.core.trafficmonitoring.TravelTimeCalculator
 import org.matsim.households.Household
 import org.matsim.utils.objectattributes.AttributeConverter
 import org.matsim.vehicles.Vehicle
-
-import scala.collection.JavaConverters._
-import scala.collection.mutable
-import scala.collection.mutable.ListBuffer
-import scala.concurrent.Await
 
 trait BeamHelper extends LazyLogging {
 
@@ -62,68 +63,6 @@ trait BeamHelper extends LazyLogging {
     | _____________________________________
     |
     """.stripMargin
-
-  private val argsParser = new scopt.OptionParser[Arguments]("beam") {
-    opt[String]("config")
-      .action(
-        (value, args) =>
-          args.copy(
-            config = Some(BeamConfigUtils.parseFileSubstitutingInputDirectory(value)),
-            configLocation = Option(value)
-        )
-      )
-      .validate(
-        value =>
-          if (value.trim.isEmpty) failure("config location cannot be empty")
-          else success
-      )
-      .text("Location of the beam config file")
-    opt[String]("cluster-type")
-      .action(
-        (value, args) =>
-          args.copy(clusterType = value.trim.toLowerCase match {
-            case "master" => Some(Master)
-            case "worker" => Some(Worker)
-            case _        => None
-          })
-      )
-      .text("If running as a cluster, specify master or worker")
-    opt[String]("node-host")
-      .action((value, args) => args.copy(nodeHost = Option(value)))
-      .validate(value => if (value.trim.isEmpty) failure("node-host cannot be empty") else success)
-      .text("Host used to run the remote actor system")
-    opt[String]("node-port")
-      .action((value, args) => args.copy(nodePort = Option(value)))
-      .validate(value => if (value.trim.isEmpty) failure("node-port cannot be empty") else success)
-      .text("Port used to run the remote actor system")
-    opt[String]("seed-address")
-      .action((value, args) => args.copy(seedAddress = Option(value)))
-      .validate(
-        value =>
-          if (value.trim.isEmpty) failure("seed-address cannot be empty")
-          else success
-      )
-      .text(
-        "Comma separated list of initial addresses used for the rest of the cluster to bootstrap"
-      )
-    opt[Boolean]("use-local-worker")
-      .action((value, args) => args.copy(useLocalWorker = Some(value)))
-      .text(
-        "Boolean determining whether to use a local worker. " +
-        "If cluster is NOT enabled this defaults to true and cannot be false. " +
-        "If cluster is specified then this defaults to false and must be explicitly set to true. " +
-        "NOTE: For cluster, this will ONLY be checked if cluster-type=master"
-      )
-
-    checkConfig(
-      args =>
-        if (args.useCluster && (args.nodeHost.isEmpty || args.nodePort.isEmpty || args.seedAddress.isEmpty))
-          failure("If using the cluster then node-host, node-port, and seed-address are required")
-        else if (args.useCluster && !args.useLocalWorker.getOrElse(true))
-          failure("If using the cluster then use-local-worker MUST be true (or unprovided)")
-        else success
-    )
-  }
 
   private def updateConfigForClusterUsing(
     parsedArgs: Arguments,
@@ -273,7 +212,7 @@ trait BeamHelper extends LazyLogging {
   }
 
   def prepareConfig(args: Array[String], isConfigArgRequired: Boolean): (Arguments, TypesafeConfig) = {
-    val parsedArgs = argsParser.parse(args, init = Arguments()) match {
+    val parsedArgs = ArgumentsParser.parseArguments(args) match {
       case Some(pArgs) => pArgs
       case None =>
         throw new IllegalArgumentException(
@@ -715,25 +654,3 @@ trait BeamHelper extends LazyLogging {
 }
 
 case class MapStringDouble(data: Map[String, Double])
-
-case class Arguments(
-  configLocation: Option[String] = None,
-  config: Option[TypesafeConfig] = None,
-  clusterType: Option[ClusterType] = None,
-  nodeHost: Option[String] = None,
-  nodePort: Option[String] = None,
-  seedAddress: Option[String] = None,
-  useLocalWorker: Option[Boolean] = None
-) {
-  val useCluster: Boolean = clusterType.isDefined
-}
-
-sealed trait ClusterType
-
-case object Master extends ClusterType {
-  override def toString = "master"
-}
-
-case object Worker extends ClusterType {
-  override def toString = "worker"
-}
